refactor: domain-based folder structure replacing pages/ui/ nesting
