--- src/pages/profile/index.ts
@@ -1,3 +1,0 @@
-export { MePage } from './ui/me-page/MePage';
-export { NotificationsPage } from './ui/notifications-page/NotificationsPage';
-export { SecurityPage } from './ui/security-page/SecurityPage';
