+++ src/pages/profile/me/index.ts
@@ -1,0 +1,1 @@
+export { MePage } from './ui/MePage';
