+++ src/pages/profile/me/ui/profile-section/ProfileSection.tsx
@@ -1,7 +1,7 @@
 import { CardSection, FloatingSaveBar } from 'shared/ui';
 import { IdentityItem } from './IdentityItem';
 import { ProfileForm } from './ProfileForm';
-import { useMePage } from '../../../model/useMePage';
+import { useMePage } from '../../model/useMePage';
 
 export function ProfileSection() {
   const { form, profile, isDirty, isPending, onSubmit, onDiscard } = useMePage();
