+++ src/pages/profile/notifications/api/useUpdateNotifications.ts
