+++ src/pages/profile/notifications/index.ts
@@ -1,0 +1,1 @@
+export { NotificationsPage } from './ui/NotificationsPage';
