+++ src/pages/profile/security/index.ts
@@ -1,0 +1,1 @@
+export { SecurityPage } from './ui/SecurityPage';
